Backend/requirements.txt
flask
pydub
whisper
torch
ffmpeg-python
openai-whisper
google-genai
python-dotenv